fixes to comments
