+++ src/components/Tokens/TokenTable/TokenRow.tsx
@@ -2,17 +2,10 @@ import { Trans } from '@lingui/macro'
 import { ParentSize } from '@visx/responsive'
 import { sendAnalyticsEvent } from 'components/AmplitudeAnalytics'
 import { EventName } from 'components/AmplitudeAnalytics/constants'
-import LineChart from 'components/Charts/LineChart'
+import SparklineChart from 'components/Charts/SparklineChart'
 import CurrencyLogo from 'components/CurrencyLogo'
 import { getChainInfo } from 'constants/chainInfo'
-import { curveCardinal, scaleLinear } from 'd3'
-import {
-  getDurationDetails,
-  PricePoint,
-  SingleTokenData,
-  TimePeriod,
-  useTokenPricesFromFragment,
-} from 'graphql/data/Token'
+import { getDurationDetails, SingleTokenData, TimePeriod } from 'graphql/data/Token'
 import { useCurrency } from 'hooks/Tokens'
 import { useAtomValue } from 'jotai/utils'
 import { ReactNode } from 'react'
@@ -39,7 +32,7 @@ import {
   useSetSortCategory,
   useToggleFavorite,
 } from '../state'
-import { DATA_EMPTY, formatDelta, getDeltaArrow, getPriceBounds } from '../TokenDetails/PriceChart'
+import { formatDelta, getDeltaArrow } from '../TokenDetails/PriceChart'
 import { Category, SortDirection } from '../types'
 import { DISPLAYS } from './TimeSelector'
 
@@ -469,7 +462,6 @@ export default function LoadedRow({
   tokenData: SingleTokenData
   timePeriod: TimePeriod
 }) {
-  const theme = useTheme()
   const tokenAddress = tokenData?.tokens?.[0].address
   const currency = useCurrency(tokenAddress)
   const tokenName = tokenData?.name
@@ -483,17 +475,6 @@ export default function LoadedRow({
   const { volume, pricePercentChange } = getDurationDetails(tokenData, timePeriod)
   const arrow = pricePercentChange ? getDeltaArrow(pricePercentChange) : null
   const formattedDelta = pricePercentChange ? formatDelta(pricePercentChange) : null
-
-  // for sparkline
-  const pricePoints = useTokenPricesFromFragment(tokenData?.prices?.[0]) ?? []
-  const hasData = pricePoints.length !== 0
-  const startingPrice = hasData ? pricePoints[0] : DATA_EMPTY
-  const endingPrice = hasData ? pricePoints[pricePoints.length - 1] : DATA_EMPTY
-  const widthScale = scaleLinear().domain([startingPrice.timestamp, endingPrice.timestamp]).range([0, 124])
-  const rdScale = scaleLinear().domain(getPriceBounds(pricePoints)).range([42, 0])
-
-  /* Default curve doesn't look good for the ALL chart */
-  const curveTension = timePeriod === TimePeriod.ALL ? 0.75 : 0.9
 
   const exploreTokenSelectedEventProperties = {
     chain_id: filterNetwork,
@@ -563,15 +544,12 @@ export default function LoadedRow({
           <SparkLine>
             <ParentSize>
               {({ width, height }) => (
-                <LineChart
-                  data={pricePoints}
-                  getX={(p: PricePoint) => widthScale(p.timestamp)}
-                  getY={(p: PricePoint) => rdScale(p.value)}
-                  curve={curveCardinal.tension(curveTension)}
-                  color={pricePercentChange && pricePercentChange < 0 ? theme.accentFailure : theme.accentSuccess}
-                  strokeWidth={2}
+                <SparklineChart
                   width={width}
                   height={height}
+                  tokenData={tokenData}
+                  pricePercentChange={pricePercentChange}
+                  timePeriod={timePeriod}
                 />
               )}
             </ParentSize>
